Moved GLSL objects, gave every OGL object an identifier and improved docs
