+++ src/main/java/joogl/gl/GLFrameBuffer.java
@@ -1,5 +1,8 @@
 package joogl.gl;
 
+import org.lwjgl.opengl.GL30;
+import org.lwjgl.opengl.GL43;
+
 /**
  * Framebuffer Objects are OpenGL Objects, which allow for the creation of user-defined Framebuffers. With them, one
  * can render to non-Default Framebuffer locations, and thus render without disturbing the main screen.
@@ -10,4 +13,7 @@ public interface GLFrameBuffer {
      * integer identifiers.
      * @return The procedural component of the framebuffer object */
     public int getFrameBuffer();
+
+    /** The OpenGL shader object identifier */
+    public static final int IDENTIFIER = GL30.GL_FRAMEBUFFER;
 }
